fix: FromHexString moved to /utils
