+++ universal/src/utils/from_string.cpp
@@ -8,7 +8,9 @@
 
 USERVER_NAMESPACE_BEGIN
 
-namespace utils::impl {
+namespace utils {
+
+namespace impl {
 
 [[noreturn]] void ThrowFromStringException(std::string_view message,
                                            std::string_view input,
@@ -18,6 +20,21 @@ namespace utils::impl {
       input, compiler::GetTypeName(resultType)));
 }
 
-}  // namespace utils::impl
+}  // namespace impl
+
+std::int64_t FromHexString(const std::string& str) {
+  std::int64_t result{};
+  try {
+    result = std::stoll(str, nullptr, 16);
+  } catch (std::logic_error& ex) {
+    throw std::runtime_error(fmt::format(
+        R"(utils::FromHexString error: Error while converting "{}" to std::int64_t)",
+        str));
+  }
+
+  return result;
+}
+
+}  // namespace utils
 
 USERVER_NAMESPACE_END
